LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-position-try-001-expected.txt


PASS Should be above the anchor when at initial scroll position
PASS Scroll down until the top edge of #anchor touches container but not overflowing
FAIL Scroll further down, making the first fallback position overflow by 1px assert_equals: Anchored element should be at the left of anchor expected 258 but got 358
FAIL Scroll back up so that both the first and second options fit. assert_equals: Anchored element should be at the left of anchor expected 258 but got 358
PASS Scroll further up, where the second option no longer fits
FAIL Scroll bottom-right to make the first two options overflow assert_equals: Anchored element should be at the right of anchor expected 307 but got 207
